Fix #102: Update embedded LDAP server Configuration classes moved into configuration package
