+++ powerauth-admin/src/main/java/io/getlime/security/app/admin/controller/IntegrationController.java
@@ -16,7 +16,7 @@
 
 package io.getlime.security.app.admin.controller;
 
-import io.getlime.security.app.admin.PowerAuthWebServiceConfiguration;
+import io.getlime.security.app.admin.configuration.PowerAuthWebServiceConfiguration;
 import io.getlime.powerauth.soap.v3.*;
 import io.getlime.security.powerauth.soap.spring.client.PowerAuthServiceClient;
 import org.springframework.beans.factory.annotation.Autowired;
